remove: ExceptionHandler 위치 common -> api 모듈로 이동
